backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pymysql
alembic
bcrypt
PyJWT
uuid
requests